initial commit of config_utils
